+++ src/RestApi/Device/Location/DeviceLocation.ts
@@ -1,0 +1,9 @@
+export interface AssignLocationToDevice {
+	deviceUid: string;
+	locationUid: string;
+}
+
+export interface UnassignLocationFromDevice {
+	deviceUid: string;
+	locationUid: string;
+}