+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bulk Operations
 - Location `create`, `list` (multiple), `get` (single), `update` and `delete`
 - Organization tag to location `assign` and `unassign`
+- Location to device `assign` and `unassign`
 
 ## [1.7.1] - 2021-12-17
 ### Fixed